amihan/requirements.txt
Django==3.0
djangorestframework==3.11.0
markdown==3.1.1
django-filter==2.2.0
djangorestframework-simplejwt==4.4.0
